Refactor: Generate Views.py to Folder Views
