+++ .idea/vn-kiet-django-training.iml
@@ -9,7 +9,9 @@
     </facet>
   </component>
   <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/img2xl_app/views" isTestSource="false" />
+    </content>
     <orderEntry type="jdk" jdkName="Python 2.7 (vn-kiet-django-training) (2)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="Google App Engine SDK" level="project" />
