+++ img2xl_app/urls.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 from django.conf.urls import url
 from django.contrib.auth import views as auth_views
-from . import views
+import views
 
 urlpatterns = [
     url(r'^register/$', views.register, name='register'),
