--- img2xl_app/views.py
@@ -1,1138 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-import traceback  # Thêm thư viện này ở đầu file
-import io
-import logging
-import csv
-import os
-import json
-from datetime import timedelta
-
-from django.utils import timezone
-from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, JsonResponse, HttpResponseForbidden
-from django.conf import settings
-from .forms import UploadFileForm, RegisterForm
-from .models import UploadedFile, ExtractedResult, UsageLog, Notification, NotificationManager
-from .services.bridge import process_and_save_extraction
-from django.urls import reverse
-from .services.sheets_export import export_to_google_sheets
-from django.views.decorators.http import require_POST
-from django.views.decorators.csrf import csrf_protect, csrf_exempt
-from .services.table_handler import TableFileHandler
-from .services.ai_extraction_images import generate_images_with_gemini
-from .services.ai_extraction_text import generate_text_with_gemini
-from .services.upload_image import upload_to_imgbb
-from django.shortcuts import render, redirect
-from django.contrib.auth import login
-from django.contrib import messages
-from django.template.loader import get_template
-from xhtml2pdf import pisa  # Dùng cho PDF
-import xlsxwriter           # Dùng cho Excel
-import io
-import json
-import re
-import random
-# Cần import thêm thư viện xử lý ảnh (hãy đảm bảo bạn đã cài 'Pillow' trong requirements.txt)
-from PIL import Image, ImageDraw, ImageFont
-from django.utils import timezone
-from django.contrib.auth.decorators import login_required
-from django.contrib.auth import update_session_auth_hash
-from django.contrib.auth.models import User
-from django.core.mail import send_mail
-from django.conf import settings
-from .models import UserProfile, Notification, NotificationManager
-from django.db.models import Count
-import logging
-
-logger = logging.getLogger(__name__)
-
-def home(request):
-    """
-    Chỉ làm nhiệm vụ hiển thị trang chủ và danh sách lịch sử.
-    Mọi hoạt động trích xuất đã chuyển sang extract_only_api.
-    """
-    user = request.user
-    recent_results = []
-
-    if user.is_authenticated():
-        # Lấy 10 kết quả gần nhất của user
-        recent_results = ExtractedResult.objects.filter(
-            user=user,
-            is_deleted=False
-        ).order_by('-created_at')[:10]
-
-    # Không còn xử lý request.method == 'POST' ở đây nữa
-    return render(request, 'home.html', {
-        'form': UploadFileForm(),
-        'recent_results': recent_results
-    })
-
-def result_detail(request, result_id):
-    result = get_object_or_404(ExtractedResult, pk=result_id)
-
-    # 1. Lấy dữ liệu từ handler (dạng list của Python)
-    raw_table_data = result.get_table(for_export=False)
-    raw_final_data = result.get_table(for_export=True)
-
-    # 2. CHUYỂN THÀNH CHUỖI JSON (Sẽ mất ký tự 'u' và biến None thành null)
-    # ensure_ascii=False giúp giữ nguyên tiếng Việt
-    table_json_str = json.dumps(raw_table_data, ensure_ascii=False)
-    final_table_json_str = json.dumps(raw_final_data, ensure_ascii=False)
-
-    # Sử dụng timezone.localtime để chuyển từ UTC sang Asia/Ho_Chi_Minh
-    local_draft_time = timezone.localtime(result.updated_at)
-    last_draft_time = local_draft_time.strftime('%d/%m/%Y %H:%M:%S')
-
-    # Thời gian bản chính (Modal)
-    last_final_time = "Chưa lưu"
-    if result.processed_at:
-        last_final_time = timezone.localtime(result.processed_at).strftime('%d/%m/%Y %H:%M:%S')
-
-    recent_results = ExtractedResult.objects.filter(
-            is_deleted=False
-    ).order_by('-created_at')
-    return render(request, 'result_detail.html', {
-        'result': result,
-        'table_json': table_json_str,
-        'final_table_json': final_table_json_str,
-        'recent_results': recent_results,
-        'last_draft_time': last_draft_time,
-        'last_final_time': last_final_time,
-    })
-
-def export_to_sheets(request, result_id):
-    result = get_object_or_404(ExtractedResult, id=result_id)
-    table = result.get_table()
-
-    if not table:
-        return HttpResponse("No table data to export.", status=400)
-
-    sheet_url, error = export_to_google_sheets(table, result.uploaded_file.filename)
-
-    if error:
-        return HttpResponse("Export failed: " + error)
-
-    # Redirect to the sheet or show link
-    return HttpResponse(
-        "<h2>Exported to Google Sheets Successfully!</h2>"
-        "<p>Open your sheet here: <a href='{url}' target='_blank'>{url}</a></p>"
-        "<br><a href='{detail_url}'>Back to result</a>".format(
-            url=sheet_url,
-            detail_url=reverse('result_detail', args=[result_id])
-        )
-    )
-
-# Xoá bảng tính ở trang home
-@login_required
-@require_POST  # Đảm bảo chỉ chấp nhận request POST để bảo mật
-def delete_result(request, result_id):
-    # 1. Tìm đối tượng bảng tính, đảm bảo đúng chủ sở hữu
-    result = get_object_or_404(ExtractedResult, id=result_id, user=request.user)
-
-    # 2. THỰC HIỆN SOFT DELETE
-    # Chúng ta KHÔNG gọi TableFileHandler(result).delete_file() ở đây nữa.
-    # File vật lý sẽ được xóa bởi script dọn dẹp sau 30 ngày.
-
-    result.is_deleted = True
-    # Hẹn giờ xóa vĩnh viễn sau 30 ngày (hoặc lấy từ cấu hình UserProfile nếu muốn)
-    result.delete_at = timezone.now() + timedelta(days=30)
-    result.save()
-
-    # 3. Thông báo cho người dùng
-    # Cập nhật nội dung thông báo để người dùng biết họ có 30 ngày để khôi phục
-    Notification.objects.create_notification(
-        user=request.user,
-        title=u"Đã chuyển vào thùng rác!",
-        message=u"Bảng tính '{}' đã được chuyển vào mục lưu trữ và sẽ bị xóa vĩnh viễn sau 30 ngày.".format(
-            result.title),
-        level='warning',
-        linked_to='/trash-bin/'
-    )
-
-    # 4. Điều hướng quay lại trang chủ hoặc trang danh sách
-    return redirect('home')
-
-
-def is_storage_full(user):
-    """
-    Hàm helper kiểm tra xem user đã đạt giới hạn 50 ảnh chưa.
-    Trả về True nếu đã đầy, False nếu còn chỗ.
-    """
-    if not user.is_authenticated():
-        return False  # Hoặc xử lý riêng cho khách (Guest)
-
-    # Đếm số file đang hoạt động (chưa bị xóa) của user
-    current_count = UploadedFile.objects.filter(user=user, is_deleted=False).count()
-    if current_count >=50:
-        # Thông báo vượt quá hạn mức lưu ảnh
-        Notification.objects.create_notification(
-            user=user,
-            title=u"Cần hành động!",
-            message=u"Vui lòng truy cập và kiểm soát dưới '{}' ảnh.".format(current_count),
-            level='warning',
-            linked_to='/documents/'
-        )
-    return current_count >= 50
-
-@require_POST
-def update_table_data(request, result_id):
-    if request.method == 'POST':
-        result = ExtractedResult.objects.get(pk=result_id, user=request.user)
-        data = json.loads(request.body)
-
-        # Lấy flag từ frontend: if is_draft=True -> lưu nháp, if False -> lưu final
-        is_draft_request = data.get('is_draft', True)
-
-        handler = TableFileHandler(result)
-        # Nếu is_draft_request là False nghĩa là người dùng bấm nút Save Changes (is_final=True)
-        success = handler.save_data(data.get('table_data'), is_final=not is_draft_request)
-
-        result.refresh_from_db()  # Lấy dữ liệu mới nhất vừa lưu vào DB
-
-        # Chuyển về giờ Việt Nam và định dạng chuỗi đầy đủ
-        local_now = timezone.localtime(result.updated_at)
-        full_time_str = local_now.strftime('%d/%m/%Y %H:%M:%S')
-
-        return JsonResponse({
-            'status': 'success',
-            'updated_at':full_time_str ,  # Trả về giờ VN
-            'is_draft': result.is_draft
-        })
-
-@csrf_protect
-def generate_ai_content(request):
-    """
-    API nhận prompt từ giao diện và trả về văn bản từ Gemini
-    """
-    if request.method == 'POST':
-        try:
-            # Parse dữ liệu JSON từ request body
-            data = json.loads(request.body)
-            prompt_text = data.get('prompt', '')
-            target_cell = data.get('cell', '')
-
-            if not prompt_text:
-                return JsonResponse({
-                    'status': 'error',
-                    'message': u'Nội dung yêu cầu không được để trống.'
-                })
-
-            # Gọi hàm từ services.py
-            # prompt_text có thể cần decode/encode nếu là tiếng Việt trong Py 2.7
-            result, error = generate_text_with_gemini(prompt_text)
-
-            if error:
-                return JsonResponse({
-                    'status': 'error',
-                    'message': error
-                })
-            return JsonResponse({
-                'status': 'success',
-                'result': result,
-                'cell': target_cell
-            })
-
-        except Exception as e:
-            return JsonResponse({
-                'status': 'error',
-                'message': str(e)
-            })
-
-    return JsonResponse({'status': 'error', 'message': 'Invalid Method'})
-
-def check_email_exists(request):
-    """Bước 3: Kiểm tra trùng lặp email (AJAX)"""
-    email = request.GET.get('email', '').lower().strip()
-
-    if not email:
-        return JsonResponse({'is_taken': False})
-
-    # Kiểm tra xem email đã tồn tại trong hệ thống chưa
-    is_taken = User.objects.filter(email=email).exists()
-
-    return JsonResponse({'is_taken': is_taken})
-
-
-def send_otp(request):
-    """Bước 5: Tạo và gửi mã OTP qua Gmail (AJAX)"""
-    if request.method == "POST":
-        try:
-            data = json.loads(request.body)
-            email = data.get('email', '').lower().strip()
-
-            if not email:
-                return JsonResponse({'success': False, 'message': u'Vui lòng nhập email.'}, status=400)
-
-            # Ràng buộc: Giới hạn tối đa 3 lần gửi mã trong 1 phiên (Session)
-            otp_count = request.session.get('otp_count', 0)
-            if otp_count >= 3:
-                return JsonResponse({
-                    'success': False,
-                    'message': u'Bạn đã hết lượt gửi mã (tối đa 3 lần).'
-                }, status=400)
-
-            # Tạo mã 4 số ngẫu nhiên
-            otp_code = str(random.randint(1000, 9999))
-
-            # Lưu mã OTP và email mục tiêu vào session để đối chiếu sau này
-            request.session['otp_code'] = otp_code
-            request.session['otp_target_email'] = email
-            request.session['otp_count'] = otp_count + 1
-            # Mã có hiệu lực trong 5 phút
-            request.session.set_expiry(300)
-
-            # Gửi mail qua SMTP Gmail
-            subject = u'[Xác thực] Mã đăng ký tài khoản của bạn'
-            message = u'Mã OTP của bạn là: {}. Vui lòng không cung cấp mã này cho bất kỳ ai.'.format(otp_code)
-
-            send_mail(subject, message, settings.EMAIL_HOST_USER, [email])
-
-            return JsonResponse({'success': True})
-
-        except Exception as e:
-            return JsonResponse({'success': False, 'message': str(e)}, status=500)
-
-    return JsonResponse({'success': False}, status=405)
-
-
-def verify_otp_ajax(request):
-    """Bước 5: Xác nhận mã OTP người dùng nhập (AJAX)"""
-    if request.method == "POST":
-        try:
-            data = json.loads(request.body)
-            user_otp = data.get('otp')
-            saved_otp = request.session.get('otp_code')
-            target_email = request.session.get('otp_target_email')
-
-            if saved_otp and user_otp == saved_otp:
-                # Đánh dấu session đã xác thực thành công cho email cụ thể này
-                request.session['is_otp_verified'] = True
-                request.session['verified_email'] = target_email
-                return JsonResponse({'success': True})
-            else:
-                return JsonResponse({'success': False, 'message': u'Mã xác nhận không đúng.'}, status=400)
-        except:
-            return JsonResponse({'success': False}, status=400)
-
-    return JsonResponse({'success': False}, status=405)
-
-
-def register(request):
-    """Bước 6: Xử lý đăng ký cuối cùng khi nhấn nút Hoàn tất"""
-    if request.method == "POST":
-        form = RegisterForm(request.POST)
-
-        # Lấy thông tin xác thực từ session
-        is_otp_verified = request.session.get('is_otp_verified', False)
-        verified_email = request.session.get('verified_email', '')
-
-        if form.is_valid():
-            # KIỂM TRA CHÉO BẢO MẬT:
-            # 1. Đã qua bước OTP chưa?
-            # 2. Email trong form có đúng là email vừa được xác thực OTP không?
-            current_email = form.cleaned_data.get('email').lower().strip()
-
-            if is_otp_verified and current_email == verified_email:
-                # Lưu User vào Auth User
-                user = form.save()
-
-                # Tạo UserProfile đi kèm (Yêu cầu của bạn)
-                UserProfile.objects.get_or_create(user=user)
-
-                # Thông báo đăng ký tài khoản thành công
-                Notification.objects.create_notification(
-                    user=user,
-                    title=u"Chào mừng thành viên mới!",
-                    # Sửa: Dùng user.username thay vì res_obj.title (vì lúc này chưa có bảng tính nào)
-                    message=u"Chào mừng '{}' đã gia nhập Extractor AI.".format(user.username),
-                    level='success',  # Sửa lỗi chính tả 'succes' -> 'success'
-                    linked_to=None
-                )
-
-                # Xóa các dấu vết xác thực trong session sau khi thành công
-                keys_to_delete = ['otp_code', 'otp_target_email', 'is_otp_verified', 'verified_email']
-                for key in keys_to_delete:
-                    if key in request.session:
-                        del request.session[key]
-
-                # Đăng nhập và chuyển hướng
-                login(request, user)
-                return redirect('home')
-    else:
-        form = RegisterForm()
-
-    return render(request, 'registration/register.html', {'form': form})
-
-def forgot_password_view(request):
-    """Hiển thị trang nhập email để lấy lại mật khẩu"""
-    return render(request, 'registration/forgot_password.html')
-
-
-@csrf_exempt  # Hoặc đảm bảo JS gửi CSRF qua Header
-def reset_password_final(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-        new_password = data.get('password')
-        email = data.get('email')
-
-        # Kiểm tra bảo mật: Session phải đã xác thực OTP thành công
-        is_verified = request.session.get('is_otp_verified', False)
-        verified_email = request.session.get('verified_email', '')
-
-        if is_verified and email.lower().strip() == verified_email.lower().strip():
-            try:
-                user = User.objects.get(email=email)
-                user.set_password(new_password)
-                user.save()
-
-                # Quan trọng: Xóa session để tránh dùng lại mã cũ
-                del request.session['is_otp_verified']
-                del request.session['otp_code']
-
-                # Thông báo đặt lại mật khẩu thành công
-                Notification.objects.create_notification(
-                    user=user,
-                    title=u"Đặt lại mật khẩu!",
-                    message=u"Đặt lại mật khẩu thành công!",
-                    level='success',
-                    linked_to=None
-                )
-
-                return JsonResponse({'success': True})
-            except User.DoesNotExist:
-                return JsonResponse({'success': False, 'message': u'Tài khoản không tồn tại.'}, status=404)
-        else:
-            return JsonResponse({'success': False, 'message': u'Chưa xác thực mã OTP.'}, status=403)
-
-    return JsonResponse({'success': False}, status=405)
-
-
-def auto_cleanup_task(request):
-    now = timezone.now()
-    in_12h = now + timedelta(hours=12)
-    trash_count = 0
-    permanent_count = 0
-    warringDel_count = 0
-
-    # --- PHẦN 1: DỌN DẸP CÁC FILE ĐÃ HẾT HẠN (QUÁ KHỨ) ---
-    expired_files = UploadedFile.objects.filter(
-        delete_at__isnull=False,
-        delete_at__lte=now
-    )
-
-    for f in expired_files:
-        if not f.is_deleted:
-            # Chuyển vào thùng rác
-            f.is_deleted = True
-            f.delete_at = now + timedelta(days=30)
-            f.save()
-            trash_count += 1
-
-        else:
-            # Xóa vĩnh viễn
-            filename_storage = f.filename
-            user_storage = f.user
-            f.delete()  # Gọi hàm xóa vật lý nếu cần trước dòng này
-            permanent_count += 1
-
-    if trash_count > 0:
-        # Thông báo cho người dùng về việc chuyển vào thùng rác
-        Notification.objects.create_notification(
-            user=request.user,
-            title=u"Ảnh đã hết hạn!",
-            message=u"Tự động dọn dẹp đã chuyển '{}' ảnh vào thùng rác và sẽ xóa vĩnh viễn sau 30 ngày.".format(trash_count),
-            level='warning'
-        )
-
-    if permanent_count > 0:
-        # Thông báo cho người dùng về việc xóa vĩnh viễn
-        Notification.objects.create_notification(
-            user=request.ser,
-            title=u"Xóa vĩnh viễn!",
-            message=u"Tự động dọn dẹp đã xoá '{}' ảnh vĩnh viễn khỏi hệ thống do hết hạn lưu trữ.".format(permanent_count),
-            level='error'
-        )
-
-    # --- PHẦN 2: TÌM VÀ CẢNH BÁO CÁC FILE SẮP HẾT HẠN (TRONG 12H TỚI) ---
-    # Lấy danh sách ảnh sắp đến hạn xóa (bao gồm cả sắp vào thùng rác và sắp xóa thật)
-    upcoming_files = UploadedFile.objects.filter(
-        delete_at__gt=now,
-        delete_at__lte=in_12h
-    ).values('user').annotate(total=Count('id'))
-
-    for entry in upcoming_files:
-        from django.contrib.auth.models import User
-        target_user = User.objects.get(id=entry['user'])
-        count = entry['total']
-        warringDel_count += 1
-
-    if warringDel_count > 0:
-        # Tạo một thông báo tổng hợp duy nhất cho mỗi user để tránh spam
-        Notification.objects.create_notification(
-            user=request.user,
-            title=u"Sắp đến hạn xóa dữ liệu!",
-            message=u"Lưu ý: Bạn có {} ảnh sẽ bị xóa hoặc chuyển vào thùng rác trong vòng 12 giờ tới.".format(warringDel_count),
-            level='info',
-            linked_to='/documents/'  # Đường dẫn đến trang quản lý ảnh của bạn
-        )
-
-    return HttpResponse(u"Đã hoàn thành dọn dẹp và gửi cảnh báo 12h.")
-
-
-def export(request, result_id):
-    result = get_object_or_404(ExtractedResult, id=result_id)
-
-    # XỬ LÝ KHI NGƯỜI DÙNG ẤN NÚT DOWNLOAD (METHOD POST)
-    if request.method == 'POST':
-        # LẤY BẢN CHỐT (Final) ĐỂ XUẤT FILE
-        final_table_data = result.get_table(for_export=True)
-
-        export_type = request.POST.get('export_type', 'xlsx')
-
-        if export_type == 'png':
-            bg_color = request.POST.get('bg_color', 'white')
-            start_cell = request.POST.get('start_cell', 'A1')
-            try:
-                num_rows = int(request.POST.get('num_rows', 5))
-                num_cols = int(request.POST.get('num_cols', 5))
-            except ValueError:
-                num_rows, num_cols = 5, 5
-
-            # Truyền final_table_data vào
-            return _export_png(result, final_table_data, bg_color, start_cell, num_rows, num_cols)
-
-        else:
-            # Truyền final_table_data vào
-            return _export_excel(result, final_table_data)
-
-    # XỬ LÝ KHI TRUY CẬP URL (METHOD GET - Xem Preview)
-    # LẤY BẢN NHÁP (Draft) ĐỂ HIỂN THỊ XEM TRƯỚC
-    draft_table_data = result.get_table(for_export=False)
-
-    preview_data = [row[:10] for row in draft_table_data[:5]]
-
-    return render(request, 'includes/export_ui.html', {
-        'result': result,
-        'preview_data': preview_data,
-        'total_rows': len(draft_table_data),
-        'total_cols': max(len(r) for r in draft_table_data) if draft_table_data else 0
-    })
-
-def _export_excel(result, table_data, start_coords=None, end_coords=None):
-    # 1. Xử lý tên file: bienlai1.png -> bienlai1
-    original_name = result.uploaded_file.filename
-    base_name = os.path.splitext(original_name)[0].replace(' ', '_')
-
-    # Dùng lại logic CSV của bạn nhưng với tên file đã sạch
-    response = HttpResponse(content_type='text/csv')
-    response.write('\xef\xbb\xbf')  # BOM cho tiếng Việt
-    response['Content-Disposition'] = 'attachment; filename="%s.csv"' % base_name
-
-    writer = csv.writer(response)
-    for row in table_data:
-        clean_row = []
-        for cell in row:
-            if cell is None:
-                cell_text = u""
-            elif not isinstance(cell, unicode):
-                cell_text = unicode(str(cell), 'utf-8', errors='ignore')
-            else:
-                cell_text = cell
-            clean_row.append(cell_text.strip().encode('utf-8'))
-        if any(clean_row):
-            writer.writerow(clean_row)
-    return response
-
-
-def _export_png(result, table_data, bg_color, start_cell, num_rows, num_cols):
-    """Xuất PNG dựa trên ô bắt đầu và kích thước vùng chọn"""
-    # 1. Xử lý tên file (bỏ đuôi cũ)
-    base_name = os.path.splitext(result.uploaded_file.filename)[0].replace(' ', '_')
-
-    # 2. Parse tọa độ ô bắt đầu (VD: A1 -> col:0, row:0)
-    def parse_start(s):
-        m = re.match(r"([A-Z]+)(\d+)", s.upper())
-        if not m: return 0, 0
-        col_str, row_str = m.groups()
-        col = 0
-        for char in col_str: col = col * 26 + (ord(char) - 64)
-        return int(row_str) - 1, col - 1
-
-    s_row, s_col = parse_start(start_cell)
-
-    # 3. Giới hạn vùng chọn (Max 30x30)
-    num_rows = min(int(num_rows), 30)
-    num_cols = min(int(num_cols), 30)
-
-    # 4. Cắt mảng dữ liệu (Slice)
-    # Lấy từ hàng s_row đến s_row + num_rows
-    sliced_data = []
-    for r in range(s_row, s_row + num_rows):
-        if r < len(table_data):
-            # Lấy từ cột s_col đến s_col + num_cols
-            row_data = table_data[r][s_col: s_col + num_cols]
-            # Nếu dòng ngắn hơn số cột yêu cầu, bù thêm ô rỗng
-            while len(row_data) < num_cols:
-                row_data.append(u"")
-            sliced_data.append(row_data)
-        else:
-            # Nếu hết dòng trong data, bù dòng rỗng
-            sliced_data.append([u""] * num_cols)
-
-    # 5. Cấu hình vẽ ảnh
-    cell_w, cell_h = 140, 45  # Tăng nhẹ size ô
-    img_w, img_h = num_cols * cell_w, num_rows * cell_h
-
-    # Màu sắc palette hiện đại
-    COLOR_BG = (255, 255, 255) if bg_color != 'black' else (28, 28, 28)
-    COLOR_STRIPE = (245, 247, 249) if bg_color != 'black' else (38, 38, 38)
-    COLOR_LINE = (220, 225, 230) if bg_color != 'black' else (60, 60, 60)
-    COLOR_TEXT = (33, 37, 41) if bg_color != 'black' else (240, 240, 240)
-
-    img = Image.new('RGB', (img_w, img_h), COLOR_BG)
-    draw = ImageDraw.Draw(img)
-
-    # Load font (Nhớ copy file .ttf vào thư mục code nhé)
-    try:
-        font = ImageFont.truetype("arial.ttf", 14)
-    except:
-        font = ImageFont.load_default()
-
-    # Vẽ dữ liệu
-    for r_idx, row in enumerate(sliced_data):
-        for c_idx, cell_value in enumerate(row):
-            x0, y0 = c_idx * cell_w, r_idx * cell_h
-            rect = [x0, y0, x0 + cell_w, y0 + cell_h]
-
-            # Vẽ nền xen kẽ (Zebra)
-            fill_color = COLOR_STRIPE if r_idx % 2 == 0 else COLOR_BG
-            draw.rectangle(rect, fill=fill_color, outline=COLOR_LINE)
-
-            # Xử lý nội dung
-            val = cell_value if isinstance(cell_value, unicode) else unicode(str(cell_value), 'utf-8', errors='ignore')
-            if len(val) > 20: val = val[:17] + "..."
-
-            # Căn giữa "thần thánh"
-            try:
-                tw, th = draw.textsize(val, font=font)
-                draw.text((x0 + (cell_w - tw) / 2, y0 + (cell_h - th) / 2), val, fill=COLOR_TEXT, font=font)
-            except:
-                # Fallback nếu textsize lỗi trên bản PIL cũ
-                draw.text((x0 + 10, y0 + 12), val, fill=COLOR_TEXT, font=font)
-
-    buf = io.BytesIO()
-    img.save(buf, format='PNG')
-    buf.seek(0)
-
-    response = HttpResponse(buf.read(), content_type='image/png')
-    response['Content-Disposition'] = 'attachment; filename="%s.png"' % base_name
-    return response
-
-@login_required
-def settings_view(request):
-    # Trả về trang settings, dữ liệu user đã có sẵn trong request.user
-    return render(request, 'settings.html')
-
-#documents.html
-@login_required
-def documents_view(request):
-    sort = request.GET.get('sort', 'recent')
-    user = request.user
-
-    # 1. Lấy danh sách bảng tính (dùng cho Sidebar 20%)
-    results_query = ExtractedResult.objects.filter(user=user, is_deleted=False)
-
-    # 2. Lấy TOÀN BỘ ảnh của user (dùng cho Gallery 80%)
-    # Điều này đảm bảo ảnh vẫn hiện dù bảng tính bị xóa
-    all_images_query = UploadedFile.objects.filter(user=user, is_deleted=False)
-
-    # Logic sắp xếp cho bảng tính
-    if sort == 'oldest':
-        results = results_query.order_by('created_at')
-        all_images = all_images_query.order_by('uploaded_at')
-    else:
-        results = results_query.order_by('-updated_at')
-        all_images = all_images_query.order_by('-uploaded_at')
-
-    return render(request, 'documents.html', {
-        'results': results,  # Dùng cho sidebar
-        'all_images': all_images,  # Dùng cho gallery ảnh
-        'extractedResult': all_images.count(),  # Đếm ảnh thay vì đếm bảng
-        'current_sort': sort
-    })
-
-@require_POST
-def create_spreadsheet_blank(request):
-    """Tạo bảng trống """
-    name = request.POST.get('name', 'Untitled Spreadsheet')
-    user = request.user
-
-    # Tạo thẳng ExtractedResult với source_file_ids rỗng
-    res_obj = ExtractedResult.objects.create(
-        user=user,
-        title=name,
-        source_file_ids=[], # Bảng trống
-        status='success',
-        is_draft=True
-    )
-
-    empty_data = [["" for _ in range(5)] for _ in range(5)]
-    handler = TableFileHandler(res_obj)
-    handler.save_data(empty_data)
-
-    # Thông báo tạo bảng tính trống thành công
-    Notification.objects.create_notification(
-        user=request.user,
-        title=u"Tạo bảng trống!",
-        message=u"Đã tạo bảng tính trống!",
-        level='success',
-        linked_to=reverse('result_detail', kwargs={'result_id': res_obj.id})
-    )
-
-    return JsonResponse({'status': 'success', 'redirect_url': reverse('result_detail', args=[res_obj.id])})
-
-
-@login_required
-@require_POST
-def update_title_api(request, result_id):
-    """API: Đổi tên bảng tính"""
-    result = get_object_or_404(ExtractedResult, id=result_id, user=request.user)
-    new_title = request.POST.get('title')
-
-    if new_title:
-        result.title = new_title
-        result.save()
-        return JsonResponse({'status': 'success'})
-
-    return JsonResponse({'status': 'error', 'message': 'Thiếu tiêu đề'}, status=400)
-
-# Xoá bảng tính ở documents/
-@login_required
-@require_POST
-def delete_result_api(request, result_id):
-    """API: Xóa mềm bảng tính"""
-    result = get_object_or_404(ExtractedResult, id=result_id, user=request.user)
-
-    # THỰC HIỆN SOFT DELETE
-    result.is_deleted = True
-    # Hẹn giờ xóa vĩnh viễn (ví dụ 30 ngày sau)
-    result.delete_at = timezone.now() + timedelta(days=30)
-    result.save()
-
-    # Thông báo
-    Notification.objects.create_notification(
-        user=request.user,
-        title=u"Đã chuyển vào thùng rác!",
-        message=u"Bảng tính '{}' đã được chuyển vào mục lưu trữ và sẽ bị xóa vĩnh viễn sau 30 ngày".format(result.title),
-        level='warning'
-    )
-
-    return JsonResponse({'status': 'success'})
-
-
-@login_required
-@require_POST
-def delete_image_api(request, img_id):
-    """API: Xóa mềm 1 ảnh (Soft Delete)"""
-    # Vẫn lấy ảnh như cũ, đảm bảo đúng chủ sở hữu
-    image = get_object_or_404(UploadedFile, id=img_id, user=request.user)
-
-    # THỰC HIỆN SOFT DELETE
-    image.is_deleted = True
-    image.delete_at = timezone.now() # Đánh dấu thời điểm xóa ngay bây giờ
-    image.save()
-
-    # Thông báo (Bạn có thể sửa lại nội dung cho chính xác hơn)
-    Notification.objects.create_notification(
-        user=request.user,
-        title=u"Đã chuyển vào thùng rác!",
-        message=u"Ảnh '{}' đã được xoá khỏi thư viện và sẽ bị xóa vĩnh viễn sau 30 ngày.".format(image.filename),
-        level='info',
-        linked_to=None
-    )
-
-    return JsonResponse({
-        'status': 'success',
-        'message': 'Image moved to trash'
-    })
-
-@login_required
-@require_POST
-def bulk_delete_images_api(request):
-    """API: Xóa mềm nhiều ảnh cùng lúc"""
-    try:
-        # 1. Lấy dữ liệu JSON từ request.body
-        data = json.loads(request.body)
-        ids = data.get('ids', [])
-
-        if not ids:
-            return JsonResponse({'status': 'error', 'message': u'Không có ID nào được cung cấp'}, status=400)
-
-        # 2. Tính toán thời gian xóa vĩnh viễn (ví dụ: 30 ngày sau)
-        scheduled_delete_at = timezone.now() + timedelta(days=30)
-
-        # 3. Thực hiện cập nhật hàng loạt (Bulk Update)
-        # Chỉ cập nhật những ảnh thuộc về user hiện tại
-        updated_count = UploadedFile.objects.filter(
-            id__in=ids,
-            user=request.user
-        ).update(
-            is_deleted=True,
-            delete_at=scheduled_delete_at
-        )
-
-        if updated_count > 0:
-            # 4. Thông báo cho người dùng
-            Notification.objects.create_notification(
-                user=request.user,
-                title=u"Đã chuyển vào thùng rác!",
-                message=u"Hệ thống đã chuyển {} ảnh vào thùng rác và sẽ xoá vĩnh viễn sau 30 ngày.".format(updated_count),
-                level='warning',
-                linked_to=None
-            )
-
-        return JsonResponse({
-            'status': 'success',
-            'count': updated_count,
-            'message': u'Đã chuyển các ảnh vào thùng rác'
-        })
-
-    except ValueError:
-        return JsonResponse({'status': 'error', 'message': u'Dữ liệu JSON không hợp lệ'}, status=400)
-    except Exception as e:
-        return JsonResponse({'status': 'error', 'message': unicode(e)}, status=500)
-
-
-def update_image_info(request):
-    if request.method == "POST":
-        img_id = request.POST.get('id')
-        new_name = request.POST.get('filename')
-        duration = request.POST.get('duration')  # 'keep', '0', '5', '60', '1440'
-
-        try:
-            from .models import UploadedFile
-            img_obj = UploadedFile.objects.get(id=img_id, user=request.user)
-
-            # 1. Cập nhật tên
-            img_obj.filename = new_name
-
-            # 2. Cập nhật thời gian xóa nếu không chọn "Giữ nguyên"
-            if duration != 'keep':
-                duration_int = int(duration)
-                if duration_int > 0:
-                    img_obj.delete_at = timezone.now() + timedelta(minutes=duration_int)
-                else:
-                    img_obj.delete_at = None  # Chuyển sang vĩnh viễn
-                # Thông báo xoá hàng loạt ảnh
-                Notification.objects.create_notification(
-                    user=request.user,
-                    title=u"Tự động xoá ảnh!",
-                    message=u"Một số ảnh đã được chỉnh thời gian tự động xoá!.",
-                    level='info',
-                    linked_to='/documents/'
-                )
-
-            img_obj.save()
-            return JsonResponse({'status': 'success'})
-
-        except Exception as e:
-            logging.error(u"Error updating image info: %s", str(e))
-            return JsonResponse({'status': 'error', 'message': str(e)}, status=400)
-
-    return JsonResponse({'status': 'error'}, status=405)
-
-def bulk_update_time(request):
-    if request.method == "POST":
-        ids = request.POST.getlist('ids[]')
-        duration_min = int(request.POST.get('duration', 0))
-        user = request.user
-
-        # 1. Tính toán thời điểm xóa mới
-        new_expiry = None
-        if duration_min > 0:
-            new_expiry = timezone.now() + timedelta(minutes=duration_min)
-
-        try:
-            from .models import UploadedFile
-            # 2. Cập nhật hàng loạt tất cả các ID thuộc về User này
-            # Lệnh .update() thực hiện 1 câu lệnh SQL duy nhất, cực kỳ tối ưu
-            UploadedFile.objects.filter(
-                id__in=ids,
-                user=user
-            ).update(delete_at=new_expiry)
-
-            # Thông báo xoá hàng loạt ảnh
-            Notification.objects.create_notification(
-                user=request.user,
-                title=u"Tự động xoá ảnh!",
-                message=u"Một số ảnh đã được chỉnh thời gian tự động xoá!.",
-                level='info',
-                linked_to='/documents/'
-            )
-
-            return JsonResponse({'status': 'success', 'count': len(ids)})
-        except Exception as e:
-            return JsonResponse({'status': 'error', 'message': str(e)}, status=400)
-
-    return JsonResponse({'status': 'error'}, status=405)
-
-
-#   settings/
-@login_required
-def update_account_settings(request):
-    if request.method == 'POST':
-        profile = request.user.profile
-
-        # 1. Cập nhật thời gian tự động xóa (ép về kiểu int)
-        auto_delete = request.POST.get('auto_delete_duration')
-        if auto_delete is not None:
-            profile.auto_delete_duration = int(auto_delete)
-
-        # 2. Cập nhật Keep EXIF
-        # Checkbox trong HTML: nếu tích sẽ gửi 'on', nếu không tích sẽ không gửi gì cả
-        # profile.keep_exif = True if request.POST.get('keep_exif') == 'on' else False
-
-        # 3. Lưu vào Database
-        profile.save()
-
-        # 4. Thông báo thành công và reload lại trang
-        messages.success(request, u"Cài đặt tài khoản của bạn đã được cập nhật thành công!")
-        return redirect('settings')  # Hoặc tên URL dẫn đến trang account của bạn
-
-    return redirect('settings')
-
-
-@login_required
-def update_profile_settings(request):
-    if request.method == 'POST':
-        profile = request.user.profile
-
-        # 1. Xử lý Upload Avatar (Nếu có file mới)
-        avatar_file = request.FILES.get('avatar')
-        if avatar_file:
-            # Đọc file sang bytes
-            image_bytes = avatar_file.read()
-            # Gọi hàm của bạn
-            new_avatar_url, error = upload_to_imgbb(image_bytes)
-
-            if new_avatar_url:
-                profile.avatar_url = new_avatar_url
-            else:
-                messages.error(request, u"Không thể upload ảnh: " + unicode(error))
-
-        # 2. Cập nhật các trường thông tin khác
-        profile.full_name = request.POST.get('full_name', '')
-        profile.website = request.POST.get('website', '')
-        profile.bio = request.POST.get('bio', '')
-        profile.is_private = True  # Luôn đóng băng theo yêu cầu
-
-        profile.save()
-        messages.success(request, u"Hồ sơ đã được cập nhật thành công!")
-        return redirect('settings')
-
-    return redirect('settings')
-
-@login_required
-def change_password(request):
-    if request.method == 'POST':
-        # 1. Lấy dữ liệu từ Form
-        old_pass = request.POST.get('old_password')
-        new_pass = request.POST.get('new_password')
-        confirm_pass = request.POST.get('confirm_password')
-        user = request.user
-
-        # 2. Kiểm tra mật khẩu cũ có đúng không
-        if not user.check_password(old_pass):
-            messages.error(request, u"Mật khẩu cũ không chính xác!")
-            return redirect('settings') # Quay lại trang settings/password
-
-        # 3. Kiểm tra 2 mật khẩu mới có khớp nhau không
-        if new_pass != confirm_pass:
-            messages.error(request, u"Hai mật khẩu mới không khớp nhau!")
-            return redirect('settings')
-
-        # 4. Kiểm tra độ dài mật khẩu (tùy chọn nhưng nên có)
-        if len(new_pass) < 6:
-            messages.error(request, u"Mật khẩu mới phải có ít nhất 6 ký tự!")
-            return redirect('settings')
-
-        # 5. Thực thi đổi mật khẩu
-        user.set_password(new_pass) # Hàm này tự động băm (hash) mật khẩu
-        user.save()
-
-        # Thông báo đặt lại mật khẩu thành công
-        Notification.objects.create_notification(
-            user=user,
-            title=u"Đặt lại mật khẩu!",
-            message=u"Đặt lại mật khẩu thành công!",
-            level='success',
-            linked_to=None
-        )
-
-        # 6. CẬP NHẬT SESSION (Rất quan trọng!)
-        # Sau khi đổi mật khẩu, Django sẽ làm mới session hash.
-        # Nếu không có dòng này, người dùng sẽ bị văng ra trang Login ngay lập tức.
-        update_session_auth_hash(request, user)
-
-        messages.success(request, u"Chúc mừng! Mật khẩu đã được thay đổi thành công.")
-        return redirect('settings')
-
-    return redirect('settings')
-
-# Notifications
-# 1. Đánh dấu một thông báo là đã đọc
-def mark_as_read(request, noti_id):
-    if request.method == "POST":
-        try:
-            noti = Notification.objects.get(id=noti_id, user=request.user)
-            noti.is_read = True
-            noti.save()
-            return JsonResponse({'success': True})
-        except:
-            return JsonResponse({'success': False}, status=400)
-
-# 2. XÓA VĨNH VIỄN một thông báo
-def delete_notification(request, noti_id):
-    if request.method == "POST":
-        try:
-            noti = Notification.objects.get(id=noti_id, user=request.user)
-            noti.delete()
-            return JsonResponse({'success': True})
-        except:
-            return JsonResponse({'success': False}, status=400)
-
-# 3. Đánh dấu tất cả là đã đọc
-def mark_all_read(request):
-    if request.method == "POST":
-        Notification.objects.filter(user=request.user, is_read=False).update(is_read=True)
-        return JsonResponse({'success': True})
-
-# 4. XÓA SẠCH tất cả thông báo
-def delete_all_notifications(request):
-    if request.method == "POST":
-        # Thay vì .update(is_deleted=True), ta dùng .delete() để xóa sạch khỏi DB
-        Notification.objects.filter(user=request.user).delete()
-        return JsonResponse({'success': True})
-
-# 5. Đảo trạng thái Đọc/Chưa đọc
-def toggle_read(request, noti_id):
-    if request.method == "POST":
-        try:
-            noti = Notification.objects.get(id=noti_id, user=request.user)
-            noti.is_read = not noti.is_read
-            noti.save()
-            return JsonResponse({'success': True, 'is_read': noti.is_read})
-        except:
-            return JsonResponse({'success': False}, status=400)
-
-# Trash Bin
-@login_required
-def trash_bin_view(request):
-    """Hiển thị trang Thùng rác"""
-    deleted_results = ExtractedResult.objects.filter(user=request.user, is_deleted=True).order_by('-delete_at')
-    deleted_files = UploadedFile.objects.filter(user=request.user, is_deleted=True).order_by('-delete_at')
-    return render(request, 'trash_bin.html', {
-        'deleted_results': deleted_results,
-        'deleted_files': deleted_files,
-    })
-
-
-@login_required
-def restore_item_api(request):
-    """
-    API DUY NHẤT xử lý khôi phục.
-    Hỗ trợ: Khôi phục 1 ảnh, nhiều ảnh, 1 bảng, nhiều bảng.
-    """
-    if request.method == 'POST':
-        item_type = request.POST.get('type')
-        ids = request.POST.getlist('ids[]')
-
-        if not ids:
-            return JsonResponse({'status': 'error', 'message': u'Chưa chọn mục nào'}, status=400)
-
-        try:
-            # Lấy số lượng mục được chọn để đưa vào thông báo
-            count = len(ids)
-            label = u""  # Để phân biệt "ảnh" hay "bảng"
-
-            if item_type == 'image':
-                # Khôi phục hàng loạt ảnh
-                UploadedFile.objects.filter(id__in=ids, user=request.user).update(
-                    is_deleted=False, delete_at=None
-                )
-                label = u"ảnh"
-
-            elif item_type == 'table':
-                # Khôi phục bảng và ảnh liên quan
-                tables = ExtractedResult.objects.filter(id__in=ids, user=request.user)
-                for table in tables:
-                    table.is_deleted = False
-                    table.delete_at = None
-                    table.save()
-
-                    # Khôi phục luôn ảnh kèm theo bảng
-                    if table.source_file_ids:
-                        UploadedFile.objects.filter(
-                            id__in=table.source_file_ids,
-                            user=request.user
-                        ).update(is_deleted=False, delete_at=None)
-
-                label = u"bảng"
-            else:
-                return JsonResponse({'status': 'error', 'message': u'Loại dữ liệu lạ'}, status=400)
-
-            # Tạo nội dung thông báo động
-            msg = u"Đã khôi phục %d %s thành công!" % (count, label)
-
-            # --- LƯU THÔNG BÁO VÀO DATABASE ---
-            Notification.objects.create_notification(
-                user=request.user,
-                title=u"Khôi phục dữ liệu",
-                message=msg,
-                level='success',
-                linked_to='/documents/'  # Bạn có thể truyền link trang chủ nếu muốn
-            )
-
-            return JsonResponse({'status': 'success', 'message': msg})
-
-        except Exception as e:
-            return JsonResponse({'status': 'error', 'message': unicode(e)}, status=500)
-
-    return JsonResponse({'status': 'error', 'message': u'Method not allowed'}, status=405)
-
-
-@login_required
-@require_POST
-def generate_ai_images(request):
-    # 1. Trích xuất dữ liệu từ request
-    is_create_new = request.POST.get('save_db') == 'true'
-    current_result_id = request.POST.get('result_id')
-    languages = request.POST.get('languages', 'all')
-    
-    # Xử lý files
-    files = request.FILES.getlist('files')
-    if not files and 'file' in request.FILES:
-        files = [request.FILES['file']]
-    
-    # Xử lý thời gian xóa tự động
-    try:
-        duration = int(request.POST.get('deleteDuration', 0))
-        expiry_date = timezone.now() + timedelta(minutes=duration) if duration > 0 else None
-    except:
-        expiry_date = None
-    
-    # 2. Gọi Service
-    try:
-        result = generate_images_with_gemini(
-            user=request.user,
-            files=files,
-            is_create_new=is_create_new,
-            current_result_id=current_result_id,
-            languages=languages,
-            expiry_date=expiry_date
-        )
-        return JsonResponse(result)
-    except Exception as e:
-        return JsonResponse({'status': 'error', 'message': str(e)}, status=500)
